backend/requirements.txt
pyjwt
boto3
stripe